Continue refactor to move art creation logic
