+++ app/src/main/java/com/copixelate/data/model/ArtModels.kt
@@ -1,6 +1,5 @@
 package com.copixelate.data.model
 
-import android.graphics.Color
 import com.copixelate.art.ArtSpace
 import com.copixelate.art.PixelGrid
 import com.copixelate.art.PixelRow
@@ -9,7 +8,6 @@ import com.copixelate.data.room.DrawingEntity
 import com.copixelate.data.room.PaletteEntity
 import com.copixelate.data.room.SpaceEntity
 import com.copixelate.data.room.SpaceEntityWithArt
-import kotlin.random.Random
 
 data class IdModel(
     val localId: Long? = null,
@@ -45,55 +43,6 @@ data class SizeModel(
 data class UpdateModel(
     val key: Int,
     val value: Int
-)
-
-
-//
-// Default Art Creation
-//
-
-private const val DEFAULT_DRAWING_WIDTH = 32
-private const val DEFAULT_DRAWING_HEIGHT = 32
-private const val DEFAULT_PALETTE_SIZE = 6
-
-fun SpaceModel.createDefaultArt(
-    width: Int = DEFAULT_DRAWING_WIDTH,
-    height: Int = DEFAULT_DRAWING_HEIGHT,
-    paletteSize: Int = DEFAULT_PALETTE_SIZE
-) = copy(
-    palette = PaletteModel().createDefaultArt(paletteSize),
-    drawing = DrawingModel().createDefaultArt(width, height, paletteSize)
-)
-
-private fun PaletteModel.createDefaultArt(
-    size: Int
-) = copy(
-    pixels = List(
-        size = size,
-        init = {
-            val rand: Int = Random(System.nanoTime()).nextInt()
-            Color.argb(
-                255,
-                Color.red(rand),
-                Color.green(rand),
-                Color.blue(rand)
-            )
-        }
-    )
-)
-
-private fun DrawingModel.createDefaultArt(
-    width: Int,
-    height: Int,
-    paletteSize: Int
-) = copy(
-    size = SizeModel(x = width, y = height),
-    pixels = List(
-        size = width * height,
-        init = { index: Int ->
-            (index / 5) % paletteSize
-        }
-    )
 )
