+++ app/src/main/java/com/copixelate/viewmodel/ArtViewModel.kt
@@ -8,11 +8,11 @@ import com.copixelate.art.ArtSpaceResult
 import com.copixelate.art.PointF
 import com.copixelate.data.model.SpaceModel
 import com.copixelate.data.model.copyFrom
-import com.copixelate.data.model.createDefaultArt
 import com.copixelate.data.model.toArtSpace
 import com.copixelate.data.model.toModel
 import com.copixelate.data.repo.ArtRepo
 import com.copixelate.data.repo.UiRepo
+import com.copixelate.ui.util.createDefaultArt
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
